Update name to editProduct
