+++ editProduct.html
@@ -46,6 +46,6 @@
 
     <!-- Link to scripts of your own -->
     <script type="module" src="./src/app.js"></script>
-    <script type="module" src="./src/addProduct.js"></script>
+    <script type="module" src="./src/editProduct.js"></script>
   </body>
 </html>
